Update product, and styles bootstrap

+++ shop/products/forms.py
@@ -1,10 +1,10 @@
 from flask_wtf.file import FileAllowed, FileField, FileRequired
-from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators, FloatField
+from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators, FloatField, DecimalField
 
 
 class Addproducts(Form):
     name = StringField('Name', [validators.DataRequired()])
-    price = FloatField('Price', [validators.DataRequired()])
+    price = DecimalField('Price', [validators.DataRequired()])
     discount = IntegerField('Discount', default=0)
     stock = IntegerField('Stock', [validators.DataRequired()])
     colors = StringField('Colors', [validators.DataRequired()])
